move outer benchmark classes
