+++ profiler/src/generator/test/storage/IgniteConceptStorageTest.java
@@ -16,10 +16,17 @@
  * along with Grakn. If not, see <http://www.gnu.org/licenses/agpl.txt>.
  */
 
-package grakn.benchmark.profiler.generator.util;
+package grakn.benchmark.profiler.generator.storage;
 
-import grakn.benchmark.profiler.generator.storage.IgniteConceptStorage;
-import grakn.core.concept.*;
+import grakn.core.concept.Attribute;
+import grakn.core.concept.AttributeType;
+import grakn.core.concept.Concept;
+import grakn.core.concept.ConceptId;
+import grakn.core.concept.EntityType;
+import grakn.core.concept.Label;
+import grakn.core.concept.RelationshipType;
+import grakn.core.concept.Thing;
+import grakn.core.concept.Type;
 import org.apache.ignite.IgniteException;
 import org.apache.ignite.Ignition;
 import org.junit.AfterClass;
@@ -32,15 +39,17 @@ import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.List;
 import java.util.stream.Collectors;
 
 import static junit.framework.TestCase.assertEquals;
+import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-
-import static org.hamcrest.Matchers.containsInAnyOrder;
 
 /**
  *
@@ -190,7 +199,7 @@ public class IgniteConceptStorageTest {
     }
 
     @Test
-    public void whenConceptIsAdded_conceptIdCanBeRetrieved()  {
+    public void whenConceptIsAdded_conceptIdCanBeRetrieved() {
         int index = 0;
         this.store.addConcept(this.conceptMocks.get(index));
         ConceptId personConceptId = this.store.getConceptId(this.entityTypeLabel, index);
@@ -230,7 +239,7 @@ public class IgniteConceptStorageTest {
         }
 
         // add 6 of 7 entities as role players too
-        for (int i = 0 ; i < 6; i++) {
+        for (int i = 0; i < 6; i++) {
             Concept conceptMock = this.conceptMocks.get(i);
             Thing thing = conceptMock.asThing();
             this.store.addRolePlayer(thing.id().toString(), thing.type().label().toString(), relTypeLabel, "somerole");
@@ -248,10 +257,10 @@ public class IgniteConceptStorageTest {
         }
 
         // ad all but the attribute and relationship
-        for (int i = 0 ; i < conceptMocks.size()-2; i++) {
+        for (int i = 0; i < conceptMocks.size() - 2; i++) {
             Concept conceptMock = this.conceptMocks.get(i);
             Thing thing = conceptMock.asThing();
-            this.store.addRolePlayer(thing.id().toString(), thing.type().label().toString(), relTypeLabel,"somerole");
+            this.store.addRolePlayer(thing.id().toString(), thing.type().label().toString(), relTypeLabel, "somerole");
         }
 
         int orphanAttributes = this.store.totalOrphanAttributes();
@@ -266,7 +275,7 @@ public class IgniteConceptStorageTest {
         }
 
         // add all but the relationship (last element)
-        for (int i = 0 ; i < conceptMocks.size()-1; i++) {
+        for (int i = 0; i < conceptMocks.size() - 1; i++) {
             Concept conceptMock = this.conceptMocks.get(i);
             Thing thing = conceptMock.asThing();
             this.store.addRolePlayer(thing.id().toString(), thing.type().label().toString(), relTypeLabel, "somerole");
@@ -284,7 +293,7 @@ public class IgniteConceptStorageTest {
         }
 
         // add all as role players
-        for (int i = 0 ; i < conceptMocks.size(); i++) {
+        for (int i = 0; i < conceptMocks.size(); i++) {
             Concept conceptMock = this.conceptMocks.get(i);
             Thing thing = conceptMock.asThing();
             this.store.addRolePlayer(thing.id().toString(), thing.type().label().toString(), relTypeLabel, "somerole");
@@ -339,7 +348,7 @@ public class IgniteConceptStorageTest {
         List<ConceptId> entitiesNotPlayingRole2 = store.getIdsNotPlayingRole(personTypeLabel, relationshipType, role2);
 
         ConceptId[] correctEntities = this.conceptMocks.subList(1, 7).stream()
-                .map(concept->concept.asThing().id())
+                .map(concept -> concept.asThing().id())
                 .collect(Collectors.toList()).toArray(new ConceptId[]{});
 
         // assert matches in any order, casting to force hamcrest to use the right
@@ -382,7 +391,7 @@ public class IgniteConceptStorageTest {
         Concept anAge = this.conceptMocks.get(7);
         String ageId = anAge.asThing().id().toString();
         String ageLabel = anAge.asThing().type().label().toString();
-        this.store.addRolePlayer(ageId, ageLabel, "@has-" + ageLabel, "@has-"+ageLabel+"-value");
+        this.store.addRolePlayer(ageId, ageLabel, "@has-" + ageLabel, "@has-" + ageLabel + "-value");
 
         int orphanAttributes = this.store.totalOrphanAttributes();
         assertEquals(0, orphanAttributes);
